refactor(switch_monitor): extract fundamental function as base class
